# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Invoices CRUD API",
    "version": "1.0.0",
    "description": "InvoicesテーブルCRUD操作API"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/Invoices": {
      "get": {
        "summary": "Invoicesテーブルの全てのレコードを取得します",
        "responses": {
          "200": {
            "description": "Retrieved list of Invoices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Invoicesテーブルにレコードを追加します",
        "requestBody": {
          "description": "Invoiceテーブルに追加するオブジェクトを設定します",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "InvoiceID": {
                    "type": "number"
                  },
                  "BillNo": {
                    "type": "string"
                  },
                  "SlipNo": {
                    "type": "string"
                  },
                  "CustomerID": {
                    "type": "string"
                  },
                  "CustomerName": {
                    "type": "string"
                  },
                  "Products": {
                    "type": "string"
                  },
                  "Number": {
                    "type": "number"
                  },
                  "UnitPrice": {
                    "type": "number"
                  },
                  "Date": {
                    "type": "string"
                  }
                },
                "required": [
                  "InvoiceID",
                  "BillNo",
                  "SlipNo",
                  "CustomerID",
                  "CustomerName",
                  "Products",
                  "Number",
                  "UnitPrice",
                  "Date"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice record created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Invoices/{Id}": {
      "get": {
        "summary": "指定したIDのレコードを取得します",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "InvoiceID"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice record retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "指定したIDのレコードを更新します",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "InvoiceID"
          }
        ],
        "requestBody": {
          "description": "更新する内容のフィールドと値を設定",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "BillNo": {
                    "type": "string"
                  },
                  "SlipNo": {
                    "type": "string"
                  },
                  "CustomerID": {
                    "type": "string"
                  },
                  "CustomerName": {
                    "type": "string"
                  },
                  "Products": {
                    "type": "string"
                  },
                  "Number": {
                    "type": "number"
                  },
                  "UnitPrice": {
                    "type": "number"
                  },
                  "Date": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice record updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "指定したIDのレコードを削除します",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "InvoiceID"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice record deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}